Parse stats and use for inquirer choices
This change splits up the stats that are returned from the eslint-stats
formatter, and uses them as an array of choices that the user can 
arrow up/down and select, instead of typing in the rule name.

A small amount of tweaking has to be done to the output of the formatter,
because it calculates all of the bar sizes based on the width of the
terminal, but inquirer adds a few characters at the start for the 
selection caret/arrow.  Unadjusted, the longest bars will wrap onto a
second line.  So, we will naively remove enough whitespace from the
longest bars to prevent that.  This means that if the two longest bars
are very close, they may end up being equal in length.  And the 
truncated bars will be out-of-proportion with the rest, since the others
are not being adjusted accordingly.  But, it’s a small enough difference
that it hopefully doesn’t cause any confusion.