Use forked version of eslint-stat
This allows us to get rid of a ton of crappy logic that was attempting
to tweak the generated `stat` output, and was not doing a great job
of it.  Instead, the forked version subtracts three columns from
`process.stdout.columns`, to account for inquirer.  It’s not something
that the main package itself would find useful, and the main package
doesn’t seem to be moving too fast, so I feel comfortable relying on
a fork.

Another side benefit is that we can use the same version of `chalk`.