Regenerate lockfile
I'm not sure why, but I started to get an error of:

`npm ERR! Cannot read property 'match' of undefined`

And apparently the only way to resolve this is to remove package-lock.json
and re-install.  The dependencies were all just updated, so I'm not
worried about this.  Probably a good thing to do anyway, to get new
transient deps that might have been locked to old versions before.