Simplify / reduce dependency on stats formatter (with tests) (#136)
In exploration for https://github.com/IanVS/eslint-nibble/issues/135, I got confused about the interactions with the "stats" formatter. It is the case that the format we receive from the `eslint-stats` is not quite compatible with inquirer, and it has to be parsed a bit. On top of that, there are a lot of things in the dependency that we do not need, including its dependencies -- lodash is unnecessary, and it also then had a transitive dependency on an old version of chalk.
<img width="1511" height="339" alt="Screenshot 2025-09-11 at 1 32 07 PM" src="https://github.com/user-attachments/assets/3659f60a-5a1d-4175-a093-584efc24a454" />

Which, given recent npm exploits as well as encouraging the running of this tool on demand via `npx`, keeping dependencies low is good.

So, first I made a really careful abstraction of the stats formatter, considering the output of the inquirer format to be part of the formatter itself by wrapping the dependency. This I was then able to test. I will admit that the tests aren't perfect, in that I can't figure out how to validate chalk output. However, they are pretty close, and short of the color, they do a lot to validate how rules are processed and output.

Once we had tests for the existing functionality in place, we were able to reproduce it in this repo. The dependency was used as a reference, and in fact, the `getBarRatio` was taken whole cloth.



---------

Co-authored-by: Ian VanSchooten <ian.vanschooten@gmail.com>