Improve detailed formatter (#143)
# Description

I've written a new detailed formatter to replace the abandoned "friendly" formatter, which fixes: https://github.com/IanVS/eslint-nibble/issues/144 and is the last piece for https://github.com/IanVS/eslint-nibble/issues/137

I've tried to carry over most of the functionality, while also merging the tsc output and make some additional cleanup. i.e. one of my concerns with the old format is that I would get lost on which message went with which code. Between the colors and the spacing, the grouping was a bit unclear.

Now, the messages are sorted by file, line, and column (vs the old formatter also sorted by severity). Each message is 2 lines of the filename+path with the line and column positions, severity, the rule name and the link (using terminal-link -- so it will print the link unless it's supported to embed the link). Also this link is fixed, using the meta url value, which fixes https://github.com/IanVS/eslint-nibble/issues/113 
The second line is the message.

Then using `@babel/code-frame`, it also uses both the start and the end values to show the full error position.

Finally, it has a summary of the *files*, similar to tsc, vs the rule links.

Added tests to support and made various updates as needed.

<img width="1183" height="771" alt="Screenshot 2025-09-26 at 11 32 34 AM" src="https://github.com/user-attachments/assets/a60d8853-5663-4dd1-8aa3-042e5b7cbcf1" />