Sync package-lock file with engine update, restore node 18+ support by removing terminal-link, update README (#146)
A divergence was introduced here between the package file and the package lock file: https://github.com/IanVS/eslint-nibble/pull/145 This PR syncs the package-lock.json file back with the new project expectations.

Additionally, to fix another engine inconsistency, we're removing `terminal-link` which at it's current version has a higher engine expectation than this project and it seems better not to depend on an "old" package.

Finally, I bumped the screenshot and some details in the README given the current state of the project